Required changed for Firebase integration. Fixed quick press of digi id button on login screen. Fixed crash condition pressing back scanning QR sending asset. Fix int/float issue with asset meta processing. Fix crash displaying amount hidden mode when default locale doesn't produce a currency symbol. Fixed issue with extra large asset images.

+++ app/src/main/java/io/digibyte/presenter/activities/models/MetaModel.java
@@ -10,7 +10,7 @@ public class MetaModel {
     boolean lockStatus;
     int numOfIssuance;
     int numOfTransfers;
-    int totalSupply;
+    float totalSupply;
     int numOfHolders;
     String issueAddress;
     public IsuanceModel metadataOfIssuence;
